7. Complex types - string and int/float to tuple

+++ 0x00-python_variable_annotations/7-to_kv.py
@@ -4,6 +4,6 @@
 from typing import Tuple, Union
 
 
-def to_kv(k: str, v: Union[float, int]) -> Tuple[str, float]:
+def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
     """function docs"""
     return (k, v ** 2)
